Move Perspective Broker logic from BuildSlave to PBSlaveProto
