+++ master/buildbot/pbmanager.py
@@ -161,7 +161,7 @@ class Dispatcher(service.Service):
         d.addCallback(done)
 
         return d
-    
+
     # ICredentialsChecker
 
     def requestAvatarId(self, creds):
